pyproject.toml
# Gesco - ERP Backend API (Cameroun / CEMAC)
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gesco"
version = "1.0.0"
description = "API backend Gesco – Système de gestion intégré multi-entreprises (Cameroun, CEMAC)"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Proprietary" }
authors = [{ name = "Gesco" }]
keywords = ["erp", "fastapi", "cameroon", "cemac", "accounting", "invoicing"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.32.1",
    "sqlalchemy>=2.0.36",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.3",
    "pydantic-settings>=2.6.1",
    "python-dotenv>=1.0.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.2.1",
    "httpx>=0.28.1",
    "python-multipart>=0.0.17",
    "email-validator>=2.2.0",
    "orjson>=3.10.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
]

[project.urls]
Documentation = "https://github.com/your-org/gesco#readme"
Repository = "https://github.com/your-org/gesco"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

# --- Pytest ---
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"
filterwarnings = ["ignore::DeprecationWarning"]

# --- Ruff (lint & format) ---
[tool.ruff]
target-version = "py312"
line-length = 100
src = ["app", "tests"]
exclude = ["alembic/versions"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = [
    "E501",      # line length (géré par formatter)
    "N805",      # __tablename__(cls) : convention SQLAlchemy
    "B008",      # Query() dans arguments FastAPI (usage standard)
    "UP042",     # str, Enum → StrEnum (compat PyEnum partagé)
    "SIM102",    # nested if → and (style)
    "SIM105",    # try/except pass → suppress (lisibilité)
    "UP046",     # Generic[T] (Pydantic)
    "UP047",     # type parameters (Pydantic)
]

[tool.ruff.format]
quote-style = "double"
